Island Level and level selector menu
Created the level selection screen so now a player will select a level, then a character then they will be taken into the game. I added the island level to the project and set up jump and landing pads to get between each of the islands. I also made a script so that when the player changes islands, that island becomes the transforms parent to avoid unwanted physics